remove old files automatically

+++ Gravlens/quasar_input.py
@@ -44,9 +44,10 @@ def sl_sys_analysis():
         args["use_ext_kappa"] = sys.argv[15]
 
         # Remove previous input files
-        print('args["inbase"]', args["inbase"])
-        os.system("rm " + args["inbase"] + "/optimize*")
-        os.system("rm " + args["inbase"] + "/datafile*")
+        if os.listdir(args["inbase"] + "/"): 
+            print('args["inbase"]', args["inbase"]+"/optimize*")
+            os.system("rm " + args["inbase"] + "/optimize*")
+            os.system("rm " + args["inbase"] + "/datafile*")
 
     args = comm.bcast(args)
 
@@ -119,15 +120,18 @@ def sl_sys_analysis():
         text_file.write("data %s\n" % data_fname)
         text_file.write("\n")
         optimized_file_nr = 1
-        if int(args["opt_explore"]) > 0:
+        if bool(int(args["opt_explore"])):
             # Explore optimization: vary all parameters within their priors
-            text_file.write("setlens 1 1\n")
-            text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
             if args["use_ext_kappa"] == "yes":
+                text_file.write("setlens 1 2\n")
+                text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
                 if len(args["ext_kappa"]) > 4:
-                    text_file.write("convrg %.4f \n" % ext_kappa[ii])
+                    text_file.write("convrg %.4f 0 0 0 0 0 0 0 0 0\n" % ext_kappa[ii])
                 else:
-                    text_file.write("convrg 0.0 \n")
+                    text_file.write("convrg 0.0 0 0 0 0 0 0 0 0 0\n")
+            else:
+                text_file.write("setlens 1 1\n")
+                text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
             text_file.write("1 0 0 1 1 1 1 0 0 1\n")
             text_file.write("seed -%d\n" % randint(0, 100))
             fit_name = args["outbase"] + "/fit%s_%d" % ("explore", ii)
@@ -142,20 +146,24 @@ def sl_sys_analysis():
             # First Optimization: at fixed shear, reoptimize galaxy mass and e/PA
             text_file.write("# 1st Opt.: add fixed shear, reoptimize galaxy mass and e/PA\n")
             text_file.write("set restart = %s\n" % args["restart_1"])
-            if 'fit_name' in locals():
+            #if 'fit_name' in locals():
+            if bool(int(args["opt_explore"])):
                 text_file.write("setlens %s.start\n" % fit_name)
                 text_file.write("changevary 1\n")
             else:
-                text_file.write("setlens 1 1\n")
-                text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
                 if args["use_ext_kappa"] == "yes":
+                    text_file.write("setlens 1 2\n")
+                    text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
                     if len(args["ext_kappa"]) > 4:
-                        text_file.write("convrg %.4f \n" % ext_kappa[ii])
+                        text_file.write("convrg %.4f 0 0 0 0 0 0 0 0 0\n" % ext_kappa[ii])
                     else:
-                        text_file.write("convrg 0.0 \n")
+                        text_file.write("convrg 0.0 0 0 0 0 0 0 0 0 0\n")
+                else:
+                    text_file.write("setlens 1 1\n")
+                    text_file.write("alpha 1.0 0.0 0.0 0.1 10.0 0.0 0.0 0.0 0.0 1.0\n")
             text_file.write("1 0 0 1 1 0 0 0 0 0\n")
             if args["use_ext_kappa"] == "yes":
-                text_file.write("0\n")
+                text_file.write("0 0 0 0 0 0 0 0 0 0\n")
             fit_name = args["outbase"] + "/fit%d_%d" % (optimized_file_nr, ii)
             text_file.write("optimize %s\n" % fit_name)
             text_file.write("\n")
@@ -168,7 +176,7 @@ def sl_sys_analysis():
             text_file.write("changevary 1\n")
             text_file.write("1 0 0 1 1 1 1 0 0 0\n")
             if args["use_ext_kappa"] == "yes":
-                text_file.write("0\n")
+                text_file.write("0 0 0 0 0 0 0 0 0 0\n")
             fit_name = args["outbase"] + "/fit%d_%d" % (optimized_file_nr, ii)
             text_file.write("varyone 1 7 -90 90 37 %s\n" % fit_name)
             text_file.write("\n")
@@ -181,7 +189,7 @@ def sl_sys_analysis():
             text_file.write("changevary 1\n")
             text_file.write("1 0 0 0 0 0 0 0 0 1\n")
             if args["use_ext_kappa"] == "yes":
-                text_file.write("0\n")
+                text_file.write("0 0 0 0 0 0 0 0 0 0\n")
             fit_name = args["outbase"] + "/fit%d_%d" % (optimized_file_nr, ii)
             text_file.write("varyone 1 10 0.5 5 37 %s\n" % fit_name)
             text_file.write("\n")
@@ -194,7 +202,7 @@ def sl_sys_analysis():
             text_file.write("changevary 1\n")
             text_file.write("1 1 1 1 1 1 1 0 0 0\n")
             if args["use_ext_kappa"] == "yes":
-                text_file.write("0\n")
+                text_file.write("0 0 0 0 0 0 0 0 0 0\n")
             fit_name = args["outbase"] + "/fit%d_%d" % (optimized_file_nr, ii)
             text_file.write("optimize %s\n" % fit_name)
             text_file.write("\n")
